Improve UX of using typed connections
We're using some type-level magic to make specifying whether
write privileges or a transaction context are required, optional.

The idea is that one uses the `TypedConnection` or `WriterConnection`
traits using "impl trait in argument position" and specifies the
associated types only when desired. This leads to more readable code
with less of the type-level machinery to do the static checks exposed.