Removes the initial skipping of messages by the Consumer until an Init message is found
Reason: There should never be another message, so if there is, it means
the client is talking to an incompatible server.
In such a scenario (which we've now seen in the wild), we'd rather have
the client violently complain rather than silently hang. That makes
detecting this failure scenario much faster and simpler.