Clarify consumer log message when chunk reservation expired
The original message was a "todo"/future idea from when opsqueue was
more actively being developed. It has a high severity and it sounds
actionable to users reading the logs, whereas there isn't really any
way they can change the handling. This is a little misleading, so this
commit changes the message to be clearer about what's going on.